Android 16.0.0 release 4

* tag 'android-16.0.0_r4': (6818 commits)
  Fix skip epoll flag behavior
  Fix predictive back system animation stuck
  notifyFinished: Send requestedVisibleTypes before update animatingTypes
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Do not select displays that cannot host the activities as fallback
  Add a check for enable_display_content_mode_management flag when checking display.canHostTasks()
  Don't allow camera gesture while SIM locked
  Remove bright red opaque status bar.
  Reorder EXECUTE_APP_FUNCTIONS permission declaration
  Validate displayName for AssociationRequest
  ...

+++ services/tests/servicestests/src/com/android/server/net/NetworkPolicyManagerServiceTest.java
@@ -643,8 +643,10 @@ public class NetworkPolicyManagerServiceTest {
 
     @After
     public void removeFiles() throws Exception {
-        for (File file : mPolicyDir.listFiles()) {
-            file.delete();
+        if (mPolicyDir != null && mPolicyDir.exists()) {
+            for (File file : mPolicyDir.listFiles()) {
+                file.delete();
+            }
         }
     }
 
@@ -3025,6 +3027,7 @@ public class NetworkPolicyManagerServiceTest {
      * Creates a mock and registers it to {@link LocalServices}.
      */
     private static <T> T addLocalServiceMock(Class<T> clazz) {
+        LocalServices.removeServiceForTest(clazz);
         final T mock = mock(clazz);
         LocalServices.addService(clazz, mock);
         return mock;
